Documentation and a couple tweaks. (#25)
* Some docs & cleanup

* More docs, and (hopefully) make other alternatives work.

* Mo'Docs

* Fix compile error and tell travis to hit the correct gradle target.
