+++ .travis.yml
@@ -7,7 +7,10 @@ os:
   - linux
 
 before_install:
-- chmod +x gradlew
+  - chmod +x gradlew
+
+install:
+  - ./gradlew build
 
 cache:
   directories:
