+++ src/main/java/org/millenaire/CommonUtilities.java
@@ -4,6 +4,8 @@ import net.minecraft.block.Block;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.Blocks;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.BlockPos;
+import net.minecraft.util.EnumFacing;
 import org.millenaire.gui.MillAchievement;
 import org.millenaire.items.MillItems;
 
@@ -13,9 +15,9 @@ public class CommonUtilities {
     public static Random random = new Random();
 
     /**
-     * pretty much orgainizes the player's money
+     * Sorts out a player's money, converting any stacks of smaller coins into one of the next denomination.
      *
-     * @param playerIn The player to orgainize
+     * @param playerIn The player to organize money for.
      */
     public static void changeMoney(EntityPlayer playerIn) {
         ItemStack denier = new ItemStack(MillItems.denier, 0, 0);
@@ -62,29 +64,29 @@ public class CommonUtilities {
     }
 
     /**
-     * yep
+     * Gets a random number.
      *
-     * @return A random non-zero integer
+     * @return A random decimal between 0.1 and 1.1
      */
     public static float getRandomNonzero() {
         return random.nextFloat() + 0.1f;
     }
 
     /**
-     * gets a random Millager Gender
+     * Gets a random Villager Gender
      *
-     * @return
+     * @return A random number from -2 to 0
      */
     public static int randomizeGender() {
         return random.nextInt(3) - 2;
     }
 
     /**
-     * yep
+     * Gets the appropriate block to make surfaces out of
      *
-     * @param b       the block to check
+     * @param b       The block to base the check on
      * @param surface if the ground is on the top of the ground (true) or underground (false)
-     * @return
+     * @return The surface block
      */
     public static Block getValidGroundBlock(final Block b, final boolean surface) {
         if (b == Blocks.bedrock || b == Blocks.dirt ||
@@ -109,5 +111,45 @@ public class CommonUtilities {
         }
 
         return null;
+    }
+
+    /**
+     * Convenience method, same as {@link CommonUtilities#adjustForOrientation(int, int, int, int, int, EnumFacing)} but
+     * with a BlockPos instead of individual coordinates.
+     *
+     * @param pos         The base position to add to.
+     * @param xoffset     The amount to increase the x coordinate by (may be negative)
+     * @param zoffset     The amount to increase the z coordinate by (may be negative)
+     * @param orientation The direction this object is facing.
+     * @return A BlockPos containing the adjusted coordinates.
+     */
+    public static BlockPos adjustForOrientation(final BlockPos pos, final int xoffset, final int zoffset, final EnumFacing orientation) {
+        return adjustForOrientation(pos.getX(), pos.getY(), pos.getZ(), xoffset, zoffset, orientation);
+    }
+
+    /**
+     * Adds the specified x and z offsets to the given x and z coordinates, based on which way the object is facing.
+     *
+     * @param x           The X Pos
+     * @param y           The Y Pos
+     * @param z           The Z Pos
+     * @param xoffset     The amount to increase the x coordinate by (may be negative)
+     * @param zoffset     The amount to increase the z coordinate by (may be negative)
+     * @param orientation The direction this object is facing.
+     * @return A BlockPos containing the adjusted coordinates.
+     */
+    public static BlockPos adjustForOrientation(final int x, final int y, final int z, final int xoffset, final int zoffset, final EnumFacing orientation) {
+        BlockPos pos = new BlockPos(x, y, z);
+        if (orientation == EnumFacing.SOUTH) {
+            pos = new BlockPos(x + xoffset, y, z + zoffset);
+        } else if (orientation == EnumFacing.WEST) {
+            pos = new BlockPos(x + zoffset, y, z - xoffset);
+        } else if (orientation == EnumFacing.NORTH) {
+            pos = new BlockPos(x - xoffset, y, z - zoffset);
+        } else if (orientation == EnumFacing.EAST) {
+            pos = new BlockPos(x - zoffset, y, z + xoffset);
+        }
+
+        return pos;
     }
 }
